+++ src/nrniv/nrncore_write/callbacks/nrncore_callbacks.h
@@ -39,7 +39,7 @@ void nrnthread_group_ids(int* groupids);
 int nrnthread_dat1(int tid,
                    int& n_presyn,
                    int& n_netcon,
-                   std::vector<int>& output_gid,
+                   std::vector<long long>& output_gid,
                    int*& netcon_srcgid,
                    std::vector<int>& netcon_negsrcgid_tid);
 int nrnthread_dat2_1(int tid,
